UM: Add Configuration variable DefaultPixelFormat
The EGL pixel format is set to ARGB8888 as now if this variable is not set.
The user can use this configuration variable to update the default
EGL pixel format to RGB565, ARGB1555 or ARGB4444 in powervr.ini

For AM335x Beagle Bone Black EVM, the variable should be set to RGB565.

Signed-off-by: Eric Ruei <e-ruei1@ti.com>